+++ frontend/src/main/scala/com/neo/sk/tank/front/tankClient/BrickClientImpl.scala
@@ -11,7 +11,7 @@ import org.scalajs.dom.ext.Color
   * Created by hongruying on 2018/7/10
   */
 class BrickClientImpl(
-                       override val oId: Long,
+                       override val oId: Int,
                        override protected var curBlood: Int,
                        override var position: model.Point
                      ) extends Brick{
